refactor(engineer_loop/types): split tests to new sibling tests_types_extra.rs (#1266)
Existing tests_types.rs (309 LOC) and tests_types_inline.rs (228 LOC) were
already siblings; created NEW tests_types_extra.rs for the inline-mod tests.

- src/engineer_loop/types.rs: 662 → 393 LOC
- New src/engineer_loop/tests_types_extra.rs (266 LOC)

Pure refactor: 35 tests pass.
