+++ src/engineer_loop/mod.rs
@@ -18,6 +18,8 @@ mod tests_selection_extra;
 #[cfg(test)]
 mod tests_types;
 #[cfg(test)]
+mod tests_types_extra;
+#[cfg(test)]
 mod tests_types_inline;
 #[cfg(test)]
 mod tests_verification;
